Delete Profile Picture on Account delete

+++ beMeal/Server/Authentication/deleteAccount.js
@@ -7,18 +7,36 @@ export const deleteAccount = async (req, res) => {
     const { userID } = req.body;
 
     // Find account by userName
-    const account = await Account.findOne({ userID });
+    const account = await Account.findOne({ '_id': userID });
+
     // if account does not exist, return error
     if (!account) {
       return res.status(400).json({ error: "Account does not exist" });
     }
 
-    await cloudinary.uploader.destroy(account._id.toString());
+    console.log(account._id.toString());
+    console.log(userID);
 
-    await User.deleteOne({'userID': account._id});
-    res.status(200).json({ message: "User deleted" });
+    const delImage = await cloudinary.uploader.destroy("profile_pictures/" + account._id.toString());
 
-    await Account.deleteOne({ userID });
+    if (delImage.result === 'not found') {
+      return res.status(404).json({ error: "image not deleted" });
+    }
+
+    console.log("image deleted")
+
+    const userDeleted = await User.deleteOne({'userID': account._id});
+
+    console.log('User deleted');
+    if (!userDeleted){
+      return res.status(404).json({ error: "user does not exist" });
+    }
+
+    const accountDeleted = await Account.deleteOne({ userID });
+
+    if (!accountDeleted){
+      return res.status(404).json({ error: "Account does not exist" });
+    }
     res.status(200).json({ message: "Account deleted" }); 
 
     } catch (error) {
